[WNMGDS-969] Typescript Dialog Conversion (#1462)
* first commit, functional switchover

* updated refs, included button types, functional

* small cleanup

* these shouldn't be here

* putting those back

* proper exporting in typescript

* adding ariamodalprops definition for modal

* checking if prettier hits this

* wrapping analytics test event for use with fuctional component

* fixed ordering
