--- packages/design-system/src/components/Dialog/index.js
@@ -1,1 +1,0 @@
-export { default as Dialog } from './Dialog';
